// src/models/Employee.ts
export interface IEmployee {
  listByClient: {
    id: number;
    id_cliente: number;
    cod_colaborador?: string;
    nome: string;
    idade?: number;
    data_nascimento?: Date;
    data_admissao?: Date;
    cep?: string;
    endereco?: string;
    bairro?: string;
    cidade?: string;
    estado?: string;
    email?: string;
    telefone?: string;
    cargo?: number;
    status?: string;
    setor_executante?: string;
    id_equipe?: number;
    filiais?: string;
    valor_hora?: number;
    //imagem?: Buffer;
    observacoes?: string;
  };
  listByClientId: {
    id: number;
    id_cliente: number;
    cod_colaborador?: string;
    nome: string;
    idade?: number;
    data_nascimento?: Date;
    data_admissao?: Date;
    cep?: string;
    endereco?: string;
    bairro?: string;
    cidade?: string;
    estado?: string;
    email?: string;
    telefone?: string;
    cargo?: number;
    status?: string;
    setor_executante?: string;
    id_equipe?: number;
    filiais?: string;
    valor_hora?: number;
    //imagem?: Buffer;
    observacoes?: string;
    user: {
      login: string;
      name: string;
    }[];
  };
  findById: {
    id: number;
    id_cliente: number;
    cod_colaborador?: string;
    nome: string;
    idade?: number;
    data_nascimento?: Date;
    data_admissao?: Date;
    cep?: string;
    endereco?: string;
    bairro?: string;
    cidade?: string;
    estado?: string;
    email?: string;
    telefone?: string;
    cargo?: number;
    status?: string;
    setor_executante?: string;
    id_equipe?: number;
    filiais?: string;
    valor_hora?: number;
    //imagem?: Buffer;
    observacoes?: string;
  };
}